+++ XMarsProjsct/src/XMars/Learning/Common/FeatureImpInfo.java
@@ -2,9 +2,9 @@ package XMars.Learning.Common;
 
 public class FeatureImpInfo implements java.lang.Comparable<FeatureImpInfo>
 {
-	public int FeatureId;
+	public long FeatureId;
 	public int ImpCount;
-	public FeatureImpInfo(int featureId,int impCount)
+	public FeatureImpInfo(long featureId,int impCount)
 	{
 		FeatureId = featureId;
 		ImpCount = impCount;
